+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.6.7] - Unreleased
+
+### Changed
+
+- When a context is not for a `.dts` file, diagnostic errors not related to syntax are
+  downgraded set to warnings
+
 ## [0.6.6] - 2025-11-02
 
 ### Fixed